feature: loading调优，Nprogress调优。
