+++ web/src/main.js
@@ -17,27 +17,7 @@ import { store } from '@/pinia'
 import App from './App.vue'
 import { initDom } from './utils/positionToCode'
 
-
 initDom()
-/**
- * @description 导入加载进度条，防止首屏加载时间过长，用户等待
- *
- * */
-import Nprogress from 'nprogress'
-import 'nprogress/nprogress.css'
-Nprogress.configure({ showSpinner: false, ease: 'ease', speed: 500 })
-Nprogress.start()
-
-/**
- * 无需在这块结束，会在路由中间件中结束此块内容
- * */
-
-const removeLoading = () => {
-    const element = document.getElementById('loading-box');
-    if (element) {
-       element.remove();
-    }
-}
 
 const app = createApp(App)
 app.config.productionTip = false
@@ -48,7 +28,4 @@ app
     .use(auth)
     .use(router)
     .mount('#app')
-
-removeLoading()
-
 export default app
