+++ web/src/permission.js
@@ -3,6 +3,8 @@ import { useRouterStore } from '@/pinia/modules/router'
 import getPageTitle from '@/utils/page'
 import router from '@/router'
 import Nprogress from 'nprogress'
+import 'nprogress/nprogress.css'
+Nprogress.configure({ showSpinner: false, ease: 'ease', speed: 500 })
 
 const whiteList = ['Login', 'Init']
 
@@ -110,13 +112,23 @@ router.beforeEach(async(to, from) => {
   }
 })
 
+const removeLoading = () => {
+  const element = document.getElementById('gva-loading-box');
+  if (element) {
+    element.remove();
+  }
+}
+
 router.afterEach(() => {
   // 路由加载完成后关闭进度条
   document.getElementsByClassName('main-cont main-right')[0]?.scrollTo(0, 0)
+  removeLoading()
   Nprogress.done()
 })
 
 router.onError(() => {
   // 路由发生错误后销毁进度条
+  removeLoading()
   Nprogress.remove()
 })
+
